requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.4
django-heroku==0.3.1
django-vanilla-views==1.0.5
django-widget-tweaks==1.4.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
urllib3==1.22
whitenoise==3.3.1
